Gut thread locks. Threads now poll the queue at a predictable interval
The primary reason we did this is that the old approach relied on
unsetting OMP locks that were already unset. This is undefined behavior.
GCC ignores it, but Clang segfaults. In any case, if I knew about this,
I would have never done it.

Maybe there's a smarter algorithm, but pruning is simple, reliable, and
possibly even more efficient, since it avoids master and workers
triggering each other in a feedback loop of events. Already, the replay
flicker seems to no longer be present. Anyway, much less code is always
nice.