Add deprecated flag to command nodes.
When the deprecated flag is set, the given command and its children will
not be shown by help unless they're at the top of the retrieved tree,
and will not be present in dumpcfg. They will still be usable.

Additionally, it is up to the programmer to add a note that a given
command is deprecated in its description. This is so that they can make
clear what to do instead of using the command.