Fix nqiv saving thumbnails of thumbnails again.
This regression is likely caused by an off by one error when comparing
directly against the thumbnail root using its length (stripped of its
potential trailing slash), rather than the old way of copying to a
buffer and comparing a pointer offset. This error likely causes strncmp
to quit before comparing the whole of the root against the image path.

Fix this by comparing root against X characters of the image path
instead. We also make sure it is longer and that the part that matches
the root ends with a slash, so we know we aren't dealing with directory
with the same start to its name. This should hopefully be even more
robust than before.