Unceremoniously remove 'raw' loading options.
Don't even bother with deprecating this. Remove bloat while we can.

The SDL surface is dependent on the raw data it's backed by. Loading
them separately can cause problems where the surface is still available
and the data isn't, depending on pruning options.

At the time, it was probably done this way since they were logically
different variables and implementation details of the staged loading
were unclear.

Hypothetically, if we were going to deprecate this, we'd probably just
need to tell the parser to discard the raw option and print a message.
We should also write a test for this.