Allow nqiv to read stdin commands while running.
This follows these general steps at the beginning of master thread's
main loop:

1. Lock shared data to prevent threads from taking any new tasks. Check
   the number that are in the modified section. If it's zero then
   continue, otherwise quit early and wait for another opportunity.

2. If enabled, then use a cross-platform nonblocking input to check
   there is any data on stdin, ingest, and parse it. If necessary, a
   flag will be set to let the threads restart if thread-specific info
   is modified. There is now a distinction between quitting threads
   because of errors, and for the purpose of restarts.

Components are also refactored so to use a new variable type suitable
for communicating between threads to tell whether it is running.

A few refactors were necessary for this code, and the default config is
updated with old changes.