Add 'internal' command tree for inspecting program state.
The code is refactored accordingly such that a tree of non-deprecated
commands for printing but not modifying the program's state are
available. These are primarily intended for debugging purposes. See
`helptree internal`

Additionally, these changes better lock the thread queue by actually
locking all of its children and mark the transition to accessing
primitive or common values using generic functions and a void pointer.
The indentation of help output should be more consistent, and some
initialization/destruction related tasks are reordered, so that
information is available sooner.