Make nqivf.sh allow user stdin input while still passing a file list.
Using platform-specific code, make a temporary directory (if possible)
in the working directory. Then, in that temporary directory (if mktemp
is available), or the working directory, create a FIFO
(if FIFOs can be made) or a regular file containing the file list. If
the file already exists, increment the number at the end of its name
until an appropriate file can be found. The loop will fail after 1000
to prevent infinite recursion.

Now nqiv can read commands from that file/FIFO and more instructions
from stdin. When nqiv ends, delete the file/FIFO and the temporary
directory.