Use SDL_Cond to signal workers instead of polling.
Worker threads will now sleep and wait on a conditional variable to wake
up. After waking, the same delay algorithm is still used as before to
prevent stampeding herd issues/give time for the queue to fill some
more. It is now necessary to signal threads to wake after posting events
to them.