Add color management.
Images are converted to VIPS' built-in sRGB profile. This conversion is
done *after* the interpretation conversion to the sRGB color space. The
order doesn't seem to make too much of a difference, but keep an eye on
this.

Going out of our way to break this by converting to the built in CMYK
profile but keeping the sRGB color space seems to favor putting the
profile conversion before, but this is a contrived example intended to
break things.

Also, we use the perceptual rendering intent. Should we have an
option to change that, along with the profile?