Hopefully support every possible filename.
For the nqivf.sh helper script, we now use NUL terminators for each
record and // as the delimiter between sorting fields. The first two
(timestamps and names) are guaranteed to not have any / so can be easily
selected by sort, then removed by sed when generating the actual
command. The third (paths) is effectively selected by sorting the final
commands after the removal, since they all start the same way, followed
by the paths. Newlines in the names are then finally backslash escaped.
The realpath of each argument for find is taken so that there is a
normalized format, lacking // in path sections. Some erroneous escaping
of special characters is also fixed. (\ $ `)

To handle this format, nqiv is now set to strictly enforce reading and
then parsing one command at a time in the buffer. When escaped
characters are encountered (only the necessary \n \r and \) are used,
they are decoded and written to the buffer. Internally, commands are
now NUL terminated in the buffer.