Windows-related fixes.
We fixed a few compiler-related nitpicks, especially related to printf
format specifiers.

`nqiv_agetc` has been rewritten to `nqiv_stdin_agetc` for the time-being
and functions that call it specifically check whether they are reading
stdin to use it, since in Windows, stdin can be read as a special
console type in addition to a file, depending on context. stdin is so
far, the only thing we need to read asynchronously, but this may change,
later.

`nqiv_stdin_agetc` is also guaranteed to return a 32-bit integer since
we may need a (little endian) UTF-8 byte-sequence encoded therein, as
reading the console may return wide characters which we convert.

Note when reading from the Windows console, we use characters from
key-up events. This might create a few awkward cases, but at this time,
timing key-down events for repetitions is considered an overly-niche
case.