Overhaul image zooming behavior for better aspect ratio fidelity.
Should also address viewport being bigger than display in-code and size
inconsistencies. This is the actual root issue of the alpha background
checker size issue.

Consider optimizing this, since we use a loop to fix rounding errors.

We also link -lm (for math.h) now. It's a wonder we didn't need it
earlier, since we used the header. Maybe it was just a macro we used at
the time. Would it perhaps be more compatible across toolchains to link
this anyway?