Yet more zooming corrections and refactoring.
No more looping zoomrect calculations. We now multiply the biggest
quotient of the corresponding mismatched sides (if one source side is
bigger or both are smaller) against the fit ratio to scale it. It seems
to work across all images it's been tried against.

We use helper macros to replace some of the conditions.

In general, this code does less work for better results. Performance
should also be improved.