Simplify linting rules and allow more cases.
Remove `*cppcoreguidelines*` since most of the rules are only relevant to
C++ and those that aren't have aliases in other rulesets.

The following options are no longer excluded:

`cert-dcl03-c` and `misc-static-assert` The asserts
that violated these rules were either completely redundant and would never
fail, or measured things that didn't really matter, like the size of a
pointer datatype that still pointed to the same character data.

`performance-no-int-to-ptr` caused issues due to earlier infrastructural
issues with clang-tidy not being able to find glib.

`readability-duplicate-include` For the most part, this correctly
identified includes that could be removed. In a few cases, we do
redundant includes to associate them clearly with a particular `#if` and
we will disable this selectively for that case.

`readability-isolate-declaration` This only happened incidentally a
couple times and changing them is a non-issue.

`readability-suspicious-call-argument` This only happened incidentally
with a couple calls to the same procedure. Parameters within that
procedure will be rewritten as to not trigger this. It is slightly off
from the idiom of the rest of the code, but this seems like the most
elegant solution, as opposed to incidentally disabling this or renaming
variables in the calling function. More generally, something like this
*is* nice to have.